Implement pause menu fade
A simple animations which increases the render opacity of all the elements of the pause menu. With a variable to adjust the animation speed. The same animation is played in reverse when a player exits the menu. We calculate the duration of the fade animation and have a delay of that value before returning to the game.