Menu blur (#55)
* Import blur overlay asset

* Blur overlay added to menus

Also prevented multiple pause windows being opened on top of each other. We set the game to paused. And only create another widget when the game is not paused already.

* Implement pause menu fade

A simple animations which increases the render opacity of all the elements of the pause menu. With a variable to adjust the animation speed. The same animation is played in reverse when a player exits the menu. We calculate the duration of the fade animation and have a delay of that value before returning to the game.

* Implement fade animation for main menu

Add the fade in animation for the main menu. Also added a background blur. So that the UI is distinguishable from the actual game.